requirements.txt
pymongo~=3.12.1
requests~=2.26.0
web.py~=0.62
msrest~=0.6.21
Pillow~=8.3.2
wget~=3.2
python-dateutil~=2.8.2
